[BE]  코치 목록 조회시 코치가 면담 신청이 가능한지도 같이 반환한다. (#395)
* [FE] feat: UI/UX 개선 및 버그 개선한다.  (#373)

* fix: authorization query caching 버그 해결

* fix: nickname 유무에 따른 리다이렉트 설정


* fix: Date 생성자 관련 브라우저 대응

* style: 모달창 반응형 개선

* feat: InterviewApplyPage - 코치 및 날짜 선택 값 보여주기

* refactor: TitleBox - prop -> children 변경

* feat: InterviewApplyPage - 안내 문구 텍스트 표시

* refactor: 네임명 변경 및 스타일 수정

- DisplayClickedText -> EmphasizedText



* fix: 터놓고 코치 이니셜라이저에서 제거

* [FE] fix: QA중 발생한 버그를 고친다.   (#375)

* chore: EmphasizedText - (name -> nickname)으로 변경

* style: Calendar - 모바일 대응

* [FE] feat: QA중에 나온 버그를 수정한다. (#379)

* fix: InterviewApplyPage - 다른 코치를 선택하고 되는 시간 조회할 때 원래 신청했던 시간 나오는 버그

* fix: 코멘트 버튼 텍스트 수정

* chore: CoachInterviewCreatePage - 승인하기 -> 스케쥴 생성으로 워딩 변경

* style: Time - gap 제거 및 border 겹침 제거

* fix: dayType(disable) 버그 해결

* style: Time - active 상태시 border 스타일 이슈 해결

* feat: 달력 월별 이동 기능 추가

* style: TitleBox - 스타일 의존성 제거

* refactor: 주석 제거



* fix: databaseInitializer Profile dev 삭제

* [BE] fix: CalendarResponse의 interviewStatus 필드를 json으로 변환할때 필드명을 status 를 사용하도록 수정 (#385)

* [BE] fix: interivew 수정시 특정되는시간이 간헐적으로 여러개가 보이는 이슈 해결 (#388)

* [BE] refactor: Slack Alarm WebClient -> RestTemplate 변경  (#386)

* refactor: Slack Alarm 상황별 분리

* refactor: Webclient -> RestTemplate 수정

* refactor: 알람 패키지 분리

* refactor: 테스트 yml 파일 수정

* refactor: 체크스타일 수정

* refactor: 체크스타일 수정

* refactor: 체크스타일 수정

* refactor: 불필요 파일 삭제

* [BE] fix: restDocs 404 error를 해결한다.  (#381)

* fix: WebConfig addResourceHandlers 추가

* fix: crew 내 정보 조회, 수정 http request, response 경로 수정

* [BE] chore: gson 추가 (#393)

* [BE] refactor: interviewStatus COMPLETE -> COMPLETED 변경 (#394)

* feat: 코치 목록 조회시 코치가 면담 신청이 가능한지도 같이 반환한다.

+++ backend/src/test/java/com/woowacourse/ternoko/support/fixture/InterviewFixture.java
@@ -19,9 +19,9 @@ public class InterviewFixture {
 
     public static final Interview INTERVIEW = new Interview(1L, LocalDateTime.now().plusDays(10),
             LocalDateTime.now().minusDays(2),
-            COACH1, CREW1, List.of(createFormItem(1),
-            createFormItem(2),
-            createFormItem(3)), InterviewStatusType.FIXED);
+            COACH1, CREW1, List.of(createFormItem(1L),
+            createFormItem(2L),
+            createFormItem(3L)), InterviewStatusType.FIXED);
 
     private static final LocalDateTime NOW = LocalDateTime.now().withNano(0).plusDays(2);
     public static final LocalDateTime AFTER_TWO_DAYS = LocalDateTime.of(NOW.getYear(), NOW.getMonthValue(),
@@ -36,15 +36,15 @@ public class InterviewFixture {
             new FormItemRequest("수정질문3", "수정답변3"));
 
 
-    public static final List<FormItem> FORM_ITEMS1 = List.of(createFormItem(1),
-            createFormItem(2),
-            createFormItem(3));
+    public static final List<FormItem> FORM_ITEMS1 = List.of(createFormItem(1L),
+            createFormItem(2L),
+            createFormItem(3L));
 
-    public static final List<FormItem> FORM_ITEMS2 = List.of(createFormItem(4),
-            createFormItem(5),
-            createFormItem(6));
+    public static final List<FormItem> FORM_ITEMS2 = List.of(createFormItem(4L),
+            createFormItem(5L),
+            createFormItem(6L));
 
-    private static FormItem createFormItem(int count) {
+    private static FormItem createFormItem(Long count) {
         return new FormItem(null, Question.from("고정질문" + count), Answer.from("고정답변" + count));
     }
